pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "puff-trigger"
version = "0.1.0"
description = "Particulate matter spike detector that plays audio clips"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135",
    "uvicorn[standard]>=0.43",
    "pydub>=0.25",
    "python-dotenv>=1.2",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
pi = [
    "enviroplus",
]
dev = [
    "pytest>=9.0",
    "httpx>=0.28",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
